backend/requirements-render.txt
# Production dependencies for Flask backend
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0

# AI/ML
google-generativeai==0.3.2
openai==1.6.1

# Database
chromadb==0.4.22
sentence-transformers==2.2.2

# WhatsApp integration
twilio==8.10.0
